+++ HISTORY.md
@@ -1,5 +1,8 @@
 ## Changelog ##
 
+### 2013-12-08 - v4.3 - (Dev time: 1 hour) ###
+- New: Ability to delete custom post type posts by categories
+
 ### 2013-11-11 - v4.2.2 - (Dev time: 1 hour) ###
 - Fix: Bug in deleting custom post types with hypen